Minor adjustments

+++ hypeskulls-minter/HSVAPOR.Minter/Model/Utxo.cs
@@ -4,7 +4,7 @@ class Utxo
 	public string TxHash { get; set; } = string.Empty;
 	public string TxIx { get; set; } = string.Empty;
 	public string Amount { get; set; } = string.Empty;
-	public uint Lovelace { get; set; }
+	public ulong Lovelace { get; set; }
 
 	public Utxo (string rawUtxo) 
 	{
@@ -17,7 +17,7 @@ class Utxo
 
 			var amounts = Amount.Split("lovelace");
 			if(amounts.Length < 1) return;
-			Lovelace = uint.Parse(amounts[0].Trim());
+			Lovelace = ulong.Parse(amounts[0].Trim());
 		}
 	}
 }
